+++ src/language-css/utils.js
@@ -1,6 +1,5 @@
 "use strict";
 
-const { isNonEmptyArray } = require("../common/util.js");
 const colorAdjusterFunctions = new Set([
   "red",
   "green",
@@ -28,7 +27,6 @@ const colorAdjusterFunctions = new Set([
   "hwb",
   "hwba",
 ]);
-const moduleRuleNames = new Set(["import", "use", "forward"]);
 
 function getAncestorCounter(path, typeOrTypes) {
   const types = Array.isArray(typeOrTypes) ? typeOrTypes : [typeOrTypes];
@@ -58,46 +56,6 @@ function getPropOfDeclNode(path) {
     declAncestorNode.prop &&
     declAncestorNode.prop.toLowerCase()
   );
-}
-
-function hasSCSSInterpolation(groupList) {
-  if (isNonEmptyArray(groupList)) {
-    for (let i = groupList.length - 1; i > 0; i--) {
-      // If we find `#{`, return true.
-      if (
-        groupList[i].type === "word" &&
-        groupList[i].value === "{" &&
-        groupList[i - 1].type === "word" &&
-        groupList[i - 1].value.endsWith("#")
-      ) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-function hasStringOrFunction(groupList) {
-  if (isNonEmptyArray(groupList)) {
-    for (let i = 0; i < groupList.length; i++) {
-      if (groupList[i].type === "string" || groupList[i].type === "func") {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-function isSCSS(parser, text) {
-  const hasExplicitParserChoice = parser === "less" || parser === "scss";
-  const IS_POSSIBLY_SCSS = /(?:\w\s*:\s*[^:}]+|#){|@import[^\n]+(?:url|,)/;
-  return hasExplicitParserChoice
-    ? parser === "scss"
-    : IS_POSSIBLY_SCSS.test(text);
-}
-
-function isSCSSVariable(node) {
-  return Boolean(node && node.type === "word" && node.value.startsWith("$"));
 }
 
 function isWideKeywords(value) {
@@ -267,19 +225,6 @@ function isSCSSControlDirectiveNode(node) {
   );
 }
 
-function isSCSSNestedPropertyNode(node) {
-  /* istanbul ignore next */
-  if (!node.selector) {
-    return false;
-  }
-
-  return node.selector
-    .replace(/\/\*.*?\*\//, "")
-    .replace(/\/\/.*\n/, "")
-    .trim()
-    .endsWith(":");
-}
-
 function isDetachedRulesetCallNode(node) {
   return node.raws && node.raws.params && /^\(\s*\)$/.test(node.raws.params);
 }
@@ -434,42 +379,8 @@ function isColorAdjusterFuncNode(node) {
   return colorAdjusterFunctions.has(node.value.toLowerCase());
 }
 
-// TODO: only check `less` when we don't use `less` to parse `css`
-function isLessParser(options) {
-  return options.parser === "css" || options.parser === "less";
-}
-
 function lastLineHasInlineComment(text) {
   return /\/\//.test(text.split(/[\n\r]/).pop());
-}
-
-function stringifyNode(node) {
-  if (node.groups) {
-    const open = node.open && node.open.value ? node.open.value : "";
-    const groups = node.groups.reduce(
-      (previousValue, currentValue, index) =>
-        previousValue +
-        stringifyNode(currentValue) +
-        (node.groups[0].type === "comma_group" &&
-        index !== node.groups.length - 1
-          ? ","
-          : ""),
-      ""
-    );
-    const close = node.close && node.close.value ? node.close.value : "";
-
-    return open + groups + close;
-  }
-
-  const before = node.raws && node.raws.before ? node.raws.before : "";
-  const quote = node.raws && node.raws.quote ? node.raws.quote : "";
-  const atword = node.type === "atword" ? "@" : "";
-  const value = node.value ? node.value : "";
-  const unit = node.unit ? node.unit : "";
-  const group = node.group ? stringifyNode(node.group) : "";
-  const after = node.raws && node.raws.after ? node.raws.after : "";
-
-  return before + quote + atword + value + quote + unit + group + after;
 }
 
 function isAtWordPlaceholderNode(node) {
@@ -478,10 +389,6 @@ function isAtWordPlaceholderNode(node) {
     node.type === "value-atword" &&
     node.value.startsWith("prettier-placeholder-")
   );
-}
-
-function isModuleRuleName(name) {
-  return moduleRuleNames.has(name);
 }
 
 function isConfigurationNode(node, parentNode) {
@@ -522,8 +429,6 @@ module.exports = {
   getAncestorCounter,
   getAncestorNode,
   getPropOfDeclNode,
-  hasSCSSInterpolation,
-  hasStringOrFunction,
   maybeToLowerCase,
   insideValueFunctionNode,
   insideICSSRuleNode,
@@ -531,10 +436,7 @@ module.exports = {
   insideURLFunctionInImportAtRuleNode,
   isKeyframeAtRuleKeywords,
   isWideKeywords,
-  isSCSS,
-  isSCSSVariable,
   isLastNode,
-  isLessParser,
   isSCSSControlDirectiveNode,
   isDetachedRulesetDeclarationNode,
   isRelationalOperatorNode,
@@ -552,7 +454,6 @@ module.exports = {
   hasComposesNode,
   hasParensAroundNode,
   hasEmptyRawBefore,
-  isSCSSNestedPropertyNode,
   isDetachedRulesetCallNode,
   isTemplatePlaceholderNode,
   isTemplatePropNode,
@@ -570,9 +471,7 @@ module.exports = {
   isMediaAndSupportsKeywords,
   isColorAdjusterFuncNode,
   lastLineHasInlineComment,
-  stringifyNode,
   isAtWordPlaceholderNode,
-  isModuleRuleName,
   isConfigurationNode,
   isParenGroupNode,
 };
