+++ src/language-css/utils/is-SCSS-nested-property-node.js
@@ -1,0 +1,16 @@
+"use strict";
+
+function isSCSSNestedPropertyNode(node) {
+  /* istanbul ignore next */
+  if (!node.selector) {
+    return false;
+  }
+
+  return node.selector
+    .replace(/\/\*.*?\*\//, "")
+    .replace(/\/\/.*\n/, "")
+    .trim()
+    .endsWith(":");
+}
+
+module.exports = isSCSSNestedPropertyNode;
