+++ src/language-css/utils/is-SCSS-variable.js
@@ -1,0 +1,7 @@
+"use strict";
+
+function isSCSSVariable(node) {
+  return Boolean(node && node.type === "word" && node.value.startsWith("$"));
+}
+
+module.exports = isSCSSVariable;
