+++ src/language-css/utils/is-less-parser.js
@@ -1,0 +1,8 @@
+"use strict";
+
+// TODO: only check `less` when we don't use `less` to parse `css`
+function isLessParser(options) {
+  return options.parser === "css" || options.parser === "less";
+}
+
+module.exports = isLessParser;
