+++ src/language-css/utils/is-module-rule-name.js
@@ -1,0 +1,9 @@
+"use strict";
+
+const moduleRuleNames = new Set(["import", "use", "forward"]);
+
+function isModuleRuleName(name) {
+  return moduleRuleNames.has(name);
+}
+
+module.exports = isModuleRuleName;
